Add necessary jars
Somehow IntelliJ was not picking up some of the maven files. Instead of
wrestling with it, I decided to embed them here.

In particular pay attention to jsr166y which I had to compile and jar
because the version on http://g.oswego.edu/dl/concurrency-interest/
turns out to be compiled with Java 7.